Move credentials provisioning under pkg/cmd
