+++ pkg/cmd/provisioning/keypair_test.go
